More compile fixes on Windows. Actually, still under MKS Toolkit.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@1622 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ lib/sk/io/unix/test/FileDescriptorStreamsTest.cc
@@ -58,7 +58,7 @@ sk::io::test::FileDescriptorStreamsTest::
 testReadWrite()
 {
   dataOutput().writeInt(1234);
-  CPPUNIT_ASSERT_EQUAL(1234U, dataInput().readInt());
+  CPPUNIT_ASSERT_EQUAL(uint32_t(1234), dataInput().readInt());
 
   dataOutput().writeChars("Hello\n");
   CPPUNIT_ASSERT_EQUAL(sk::util::String("\"Hello\\n\""), dataInput().readLine().inspect());
